refactor: move file
The run-child-process function acts like an io function. Move it to the corresponding folder
